requirements.txt
numpy>=1.23.5,<2.3.0
scipy>=1.11,<1.14
scikit-learn>=1.3,<1.6
pandas
matplotlib
seaborn
requests
keras-tuner
tensorflow-intel==2.17.1